requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pillow>=10.3.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
pytest>=8.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
